server/ExtractorService/requirements.txt
boto3
python-pptx
PyPDF2
youtube-transcript-api